refactor:! merge async admin and vector client (#71)
