+++ tests/rbac/aio/test_admin_client_add_user.py
@@ -25,12 +25,12 @@ class add_user_test_case:
         ),
     ],
 )
-async def test_add_user(session_rbac_admin_client, test_case):
-    await session_rbac_admin_client.add_user(
+async def test_add_user(session_rbac_client, test_case):
+    await session_rbac_client.add_user(
         username=test_case.username, password=test_case.password, roles=test_case.roles
     )
 
-    result = await session_rbac_admin_client.get_user(username=test_case.username)
+    result = await session_rbac_client.get_user(username=test_case.username)
 
     assert result.username == test_case.username
 
@@ -57,12 +57,12 @@ async def test_add_user(session_rbac_admin_client, test_case):
         ),
     ],
 )
-async def test_add_user_with_roles(session_rbac_admin_client, test_case):
-    await session_rbac_admin_client.add_user(
+async def test_add_user_with_roles(session_rbac_client, test_case):
+    await session_rbac_client.add_user(
         username=test_case.username, password=test_case.password, roles=test_case.roles
     )
 
-    result = await session_rbac_admin_client.get_user(username=test_case.username)
+    result = await session_rbac_client.get_user(username=test_case.username)
 
     assert result.username == test_case.username
