+++ tests/rbac/aio/test_admin_client_get_user.py
@@ -22,12 +22,12 @@ class get_user_test_case:
         ),
     ],
 )
-async def test_get_user(session_rbac_admin_client, test_case):
-    await session_rbac_admin_client.add_user(
+async def test_get_user(session_rbac_client, test_case):
+    await session_rbac_client.add_user(
         username=test_case.username, password=test_case.password, roles=None
     )
 
-    result = await session_rbac_admin_client.get_user(username=test_case.username)
+    result = await session_rbac_client.get_user(username=test_case.username)
 
     assert result.username == test_case.username
